% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkingClassTable.r
\name{checkingClassTable}
\alias{checkingClassTable}
\title{checkingClassTable}
\usage{
checkingClassTable(integrationTable, classTable)
}
\arguments{
\item{integrationTable}{A table with at least 3 columns, name, compo, class and mz.}

\item{classTable}{A table with three columns: class for the class to be analyzed, rtmin for the minimum of retention time, rtmax for the maximale of retention time.}
}
\description{
checkingClassTable
}
