% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkingIntegrationTable.r
\name{checkingIntegrationTable}
\alias{checkingIntegrationTable}
\title{checkingIntegrationTable}
\usage{
checkingIntegrationTable(
  integrationTable,
  duplicatedName = "name",
  byClass = TRUE
)
}
\arguments{
\item{integrationTable}{a table that must contain several columns: "compo" for the composition, "name" for the lipid species; "mz" for the theoretical exact mass/z (warnings: this exact mass has to vary depending on the mode NEG/POS of the acquisition);}

\item{duplicatedName}{can be 'name' (default) or 'compo'. Detects if duplicated in this modality exists.}
}
\description{
This function checks whether duplicates exist in the integrationTable and allows to highlight potential mistakes in the integrationTable
}
\examples{
data(integrationTable)
checkingIntegrationTable(integrationTable)
}
