% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCsvOfIntensity.r
\name{getCsvOfIntensity}
\alias{getCsvOfIntensity}
\title{getCsvOfIntensity}
\usage{
getCsvOfIntensity(res, reposave)
}
\arguments{
\item{res}{results of calculateIntensity}

\item{reposave}{repository where the .csv will be saved}
}
\description{
Saves all results obtained with calculateIntensity into several .csv files. Mandatory before running getExcelOfIntensity
}
\examples{
#'  # Example of use - to be updated according to the position of your data
 #repo=getwd()  # Here in the current working directory
 #files=list.files() # Here on all the files present in the working directory
 # integrationTable=read.table("C://integrationTable.csv")
 # integrationTable=read.table("C://integrationTable.csv")
 #res=calculateIntensity(repo, files, integrationTable,classTable)
 # reposave="D://MyResults"#Where the .csv of results will be produced
 #getCsvOfIntensity(res,reposave=reposave)
}
