% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsIntensityByMass.r
\name{lcmsIntensityByMass}
\alias{lcmsIntensityByMass}
\title{Sums intensities according to specific masses along a period of time}
\usage{
lcmsIntensityByMass(
  lcms,
  breaks = NULL,
  integrationTable = NULL,
  rt = NULL,
  mz = NULL,
  by = 5e-04,
  normalization = "none",
  spar = 0,
  agregation = "sum",
  comparisonToPeaks = FALSE
)
}
\arguments{
\item{lcms}{a lcms object stemming from lcmsRead()}

\item{breaks}{a vector of breaks}

\item{integrationTable}{table of integration whose columns are named 'name','mz','inf','sup' and 'class'}

\item{rt}{vector of size 2 containing the interval of time to be studied}

\item{mz}{vector of size 2 containing the interval of mass to be studied}

\item{by}{a numeric value. If breaks and integrationTable are NULL, breaks are automatically computed from min to max by a step defined by this parameter}

\item{normalization}{"none" for no normalization, "max" for a maximale value of the peak to 100, "sum" for a total sum of the spectra of 100}

\item{spar}{0 by default. Smoothing parameter between 0 (no smoothing) and 1}

\item{agregation}{"none" "sum" or "mean".}

\item{comparisonToPeaks}{If TRUE, the intensity value is the one of the closest mass peak contained in the integration table limits}
}
\value{
an ibm object containing a data.frame (accessible by ibmObject[[1]])
}
\description{
Sums intensities according to specific masses along a period of time
}
\examples{
#data(lcms)
#int_mass=lcmsIntensityByMass(lcms,rt=c(450,455))
#head(int_mass)
}
