% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsIntensityByTime.r
\name{lcmsIntensityByTime}
\alias{lcmsIntensityByTime}
\title{Computes the intensity by timepoint for specific mz. Required to plot a chromatogram.}
\usage{
lcmsIntensityByTime(lcms, rt = NULL, mz = NULL, timeUnits = "s")
}
\arguments{
\item{lcms}{LCMS object (from lcmsRead)}

\item{rt}{A numeric(2) or two-column matrix defining the lower and upper boundary for the retention time range/window(s) for the chromatogram(s). If a matrix is provided, a chromatogram is extracted for each row. If not specified, a chromatogram representing the full retention time range is extracted.}

\item{mz}{A numeric(2) or two-column matrix defining the mass-to-charge (mz) range(s) for the chromatogram(s). For each spectrum/retention time, all intensity values within this mz range are aggregated to result in the intensity value for the spectrum/retention time. If not specified, the full mz range is considered.}

\item{timeUnits}{"s" or "min": units of the time column}
}
\description{
Computes the intensity by timepoint for specific mz. Required to plot a chromatogram.
}
\examples{
#data(lcms)
#int_time=lcmsIntensityByTime(lcms,rt=c(450,460))
#head(int_time)
}
