% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsRead.r
\name{lcmsRead}
\alias{lcmsRead}
\title{lcmsRead}
\usage{
lcmsRead(file, rt = NULL, mz = NULL, msLevel = 1)
}
\arguments{
\item{file}{name of the file to be read}

\item{rt}{a vector of size two indicating a range of retention time to be selected}

\item{mz}{a vector of size two indicating a range of mass to be selected}

\item{msLevel}{1 for MS1, 2 for MS2 and so on...}
}
\value{
a MSnExp object if integrationTable is NULL, a dataframe ifelse
}
\description{
lcmsRead
}
\examples{
#  DO NOT RUN
#  file <- dir(system.file(package = "chemosensR", dir = "extdata"),full.name=TRUE,pattern="mzXML$")
# lcms=lcmsRead(file)
}
