% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsReadList.r
\name{lcmsReadListIBM}
\alias{lcmsReadListIBM}
\title{lcmsReadListIBM}
\usage{
lcmsReadListIBM(
  listFiles,
  normalization = "none",
  rt = NULL,
  integrationTable = NULL,
  breaks = breaks,
  by = 0.1,
  mz = NULL,
  agregation = "mean",
  comparisonToPeaks = TRUE,
  wideFormat = FALSE,
  centroided = FALSE,
  ppm = 15,
  limitIntegration = 0.1,
  byCTP = 0.001,
  higherThanNoise = 10,
  minimalIntensityForPeak = 50,
  repo = "",
  msLevel = 1
)
}
\arguments{
\item{listFiles}{list of character containing the file name to be imported}

\item{normalization}{"none" for no normalization, "max" for a maximale value of the peak to 100, "sum" for a total sum of the spectra of 100}

\item{rt}{vector of size 2 containing the interval of time to be studied}

\item{integrationTable}{table of integration whose columns are named 'name','mz','inf','sup' and 'class'}

\item{breaks}{a vector of breaks}

\item{by}{a numeric value. If breaks and integrationTable are NULL, breaks are automatically computed from min to max by a step defined by this parameter}

\item{mz}{vector of size 2 containing the interval of mass to be studied}

\item{agregation}{"none" "sum" or "mean".}

\item{comparisonToPeaks}{If TRUE, the intensity value is the one of the closest mass peak contained in the integration table limits}

\item{wideFormat}{If TRUE the returned dataframe is wide (ions in columns), ifelse, the returned dataframe is long}

\item{centroided}{TRUE by default (is the data centroided). If false, the profile mode is selected and the algorithm of peak detection is different.}

\item{ppm}{select the peak with the minimal distance within the ppm distance.}

\item{limitIntegration}{0.1 by default: window where to research peaks for each ion of integrationTable}

\item{byCTP}{intervall in mass for the sum in profile mode}

\item{higherThanNoise}{in profile (when peakPicking);select only peaks which are  higherThanNoise times higher than the noise and higher than minimalIntensityForPeak}

\item{minimalIntensityForPeak}{in profile (when peakPicking); select only peaks which are  higherThanNoise times higher than the noise and higher than minimalIntensityForPeak}
}
\description{
lcmsReadListIBM
}
