% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmsReadListITM.r
\name{lcmsReadListIBT}
\alias{lcmsReadListIBT}
\title{lcmsReadListIBT}
\usage{
lcmsReadListIBT(
  listFiles,
  rt = NULL,
  mz = NULL,
  integrationTable = NULL,
  timeUnits = "s"
)
}
\arguments{
\item{listFiles}{vector of character contaning the file names to be read}

\item{rt}{a vector of length two representing an interval of rt to be analyzed}

\item{mz}{a vector of length two representing an interval of mz to be analyzed}

\item{integrationTable}{integrationTable}

\item{timeUnits}{Unit of time "s"}
}
\description{
lcmsReadListIBT
}
