% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeWithStd.r
\name{normalizeWithStandards}
\alias{normalizeWithStandards}
\title{normalizeWithStandards}
\usage{
normalizeWithStandards(
  nameXlsx = "intokPOS.xlsx",
  nameWb = NULL,
  concentrationTable = NULL
)
}
\arguments{
\item{nameXlsx}{name of the obtained .xlsx to normalize with standards: this excel must have been computed with an integrationTable including standard (a 'std' column)}

\item{nameWb}{name of the new file to be produced (corrected with standards)}

\item{concentrationTable}{if not null, a table containing two columns: a column 'class' and a column 'concentration_mg_ml'}
}
\description{
Allows to standardize data with the standard in each class
}
