% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIntensity.r
\name{calculateIntensity}
\alias{calculateIntensity}
\title{calculateIntensity}
\usage{
calculateIntensity(
  repoData = repoData,
  files,
  integrationTable = integrationTable,
  classTable,
  nameFile = "BL.xlsx",
  centroided = T,
  ppmThreshold = 10,
  ppm = 10,
  limitIntegration = 0.1,
  repo = "",
  byCTP = 0.001,
  higherThanNoise = 10,
  minimalIntensityForPeak = 50
)
}
\arguments{
\item{integrationTable}{table of integration whose columns are named 'name','mz','inf','sup' and 'class'}

\item{centroided}{TRUE by default (is the data centroided). If false, the profile mode is selected and the algorithm of peak detection is different.}

\item{ppm}{select the peak with the minimal distance within the ppm distance.}

\item{limitIntegration}{0.1 by default: window where to research peaks for each ion of integrationTable}

\item{byCTP}{intervall in mass for the sum in profile mode}

\item{higherThanNoise}{in profile (when peakPicking);select only peaks which are  higherThanNoise times higher than the noise and higher than minimalIntensityForPeak}

\item{minimalIntensityForPeak}{in profile (when peakPicking); select only peaks which are  higherThanNoise times higher than the noise and higher than minimalIntensityForPeak}
}
\description{
calculateIntensity
}
