% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCsvOfIntensity.r
\name{getCsvOfIntensity}
\alias{getCsvOfIntensity}
\title{getCsvOfIntensity}
\usage{
getCsvOfIntensity(res, reposave)
}
\arguments{
\item{res}{results of calculateIntensity}

\item{reposave}{repository where the .csv will be saved}
}
\description{
Saves all results obtained with calculateIntensity into several .csv files. Mandatory before running getExcelOfIntensity
}
