% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExcelOfIntensity.r
\name{getExcelOfIntensity}
\alias{getExcelOfIntensity}
\title{getExcelOfIntensity}
\usage{
getExcelOfIntensity(
  reposave,
  samples = NULL,
  nameFile = NULL,
  output = "int",
  timeSd = 3,
  colors = NULL
)
}
\arguments{
\item{reposave}{directory where the results will be saved}

\item{samples}{which samples must be included in the analysis (default NULL corresponds to all samples)}

\item{nameFile}{name of the excel file containing the results}

\item{output}{'int' for intensity (absolute) results or 'pct' for data normalized by the sum of the signal per class.}

\item{timeSd}{when an intensity is higher (or lower) than avg+timeSd*sd (or -), the excel cell is colored in green (or blue)}

\item{colors}{allowing to specify a color vector for the graphs in the excel}
}
\description{
Compiles all .csv files obtained with getCsvOfIntensity of one single excel file with some graphs.
}
